Module Context_mtime.Deadline

Deadlines

Deadlines

Depends on mtime.

val k_deadline : Mtime.t Context.key

Key to access a deadline from inside the request context. This can be useful for systems concerned with timeout or cancellation of a request after a certain amount of time has passed.

val deadline : Context.t -> Mtime.t option

val deadline_exn : Context.t -> Mtime.t

raises Invalid_argument
if no deadline was defined

val with_deadline : Mtime.t -> Context.t -> Context.t

Set deadline

val expired : Context.t -> bool

Has the deadline been reached? This uses Mtime_clock.now to compare it to the stored deadline. If there is no deadline then this returns false.